+++ aws/cache.go
@@ -4,19 +4,19 @@ import (
 	"context"
 
 	"github.com/cycloidio/terracognita/errcode"
+	"github.com/cycloidio/terracognita/filter"
 	"github.com/cycloidio/terracognita/provider"
-	"github.com/cycloidio/terracognita/tag"
 	"github.com/pkg/errors"
 )
 
-func cacheLoadBalancersV2(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]provider.Resource, error) {
+func cacheLoadBalancersV2(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]provider.Resource, error) {
 	rs, err := a.cache.Get(rt)
 	if err != nil {
 		if errors.Cause(err) != errcode.ErrCacheKeyNotFound {
 			return nil, errors.WithStack(err)
 		}
 
-		rs, err = albs(ctx, a, rt, tags)
+		rs, err = albs(ctx, a, rt, filters)
 		if err != nil {
 			return nil, err
 		}
@@ -30,8 +30,8 @@ func cacheLoadBalancersV2(ctx context.Context, a *aws, rt string, tags []tag.Tag
 	return rs, nil
 }
 
-func getLoadBalancersV2Arns(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]string, error) {
-	rs, err := cacheLoadBalancersV2(ctx, a, rt, tags)
+func getLoadBalancersV2Arns(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]string, error) {
+	rs, err := cacheLoadBalancersV2(ctx, a, rt, filters)
 	if err != nil {
 		return nil, err
 	}
@@ -46,14 +46,14 @@ func getLoadBalancersV2Arns(ctx context.Context, a *aws, rt string, tags []tag.T
 	return names, nil
 }
 
-func cacheLoadBalancersV2Listeners(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]provider.Resource, error) {
+func cacheLoadBalancersV2Listeners(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]provider.Resource, error) {
 	rs, err := a.cache.Get(rt)
 	if err != nil {
 		if errors.Cause(err) != errcode.ErrCacheKeyNotFound {
 			return nil, errors.WithStack(err)
 		}
 
-		rs, err = albListeners(ctx, a, rt, tags)
+		rs, err = albListeners(ctx, a, rt, filters)
 		if err != nil {
 			return nil, err
 		}
@@ -67,8 +67,8 @@ func cacheLoadBalancersV2Listeners(ctx context.Context, a *aws, rt string, tags 
 	return rs, nil
 }
 
-func getLoadBalancersV2ListenersArns(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]string, error) {
-	rs, err := cacheLoadBalancersV2Listeners(ctx, a, rt, tags)
+func getLoadBalancersV2ListenersArns(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]string, error) {
+	rs, err := cacheLoadBalancersV2Listeners(ctx, a, rt, filters)
 	if err != nil {
 		return nil, err
 	}
@@ -83,14 +83,14 @@ func getLoadBalancersV2ListenersArns(ctx context.Context, a *aws, rt string, tag
 	return names, nil
 }
 
-func cacheIAMGroups(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]provider.Resource, error) {
+func cacheIAMGroups(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]provider.Resource, error) {
 	rs, err := a.cache.Get(rt)
 	if err != nil {
 		if errors.Cause(err) != errcode.ErrCacheKeyNotFound {
 			return nil, errors.WithStack(err)
 		}
 
-		rs, err = iamGroups(ctx, a, rt, tags)
+		rs, err = iamGroups(ctx, a, rt, filters)
 		if err != nil {
 			return nil, err
 		}
@@ -104,8 +104,8 @@ func cacheIAMGroups(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]p
 	return rs, nil
 }
 
-func getIAMGroupNames(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]string, error) {
-	rs, err := cacheIAMGroups(ctx, a, rt, tags)
+func getIAMGroupNames(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]string, error) {
+	rs, err := cacheIAMGroups(ctx, a, rt, filters)
 	if err != nil {
 		return nil, err
 	}
@@ -120,14 +120,14 @@ func getIAMGroupNames(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([
 	return names, nil
 }
 
-func cacheIAMRoles(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]provider.Resource, error) {
+func cacheIAMRoles(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]provider.Resource, error) {
 	rs, err := a.cache.Get(rt)
 	if err != nil {
 		if errors.Cause(err) != errcode.ErrCacheKeyNotFound {
 			return nil, errors.WithStack(err)
 		}
 
-		rs, err = iamRoles(ctx, a, rt, tags)
+		rs, err = iamRoles(ctx, a, rt, filters)
 		if err != nil {
 			return nil, err
 		}
@@ -141,8 +141,8 @@ func cacheIAMRoles(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]pr
 	return rs, nil
 }
 
-func getIAMRoleNames(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]string, error) {
-	rs, err := cacheIAMRoles(ctx, a, rt, tags)
+func getIAMRoleNames(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]string, error) {
+	rs, err := cacheIAMRoles(ctx, a, rt, filters)
 	if err != nil {
 		return nil, err
 	}
@@ -157,14 +157,14 @@ func getIAMRoleNames(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]
 	return names, nil
 }
 
-func cacheIAMUsers(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]provider.Resource, error) {
+func cacheIAMUsers(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]provider.Resource, error) {
 	rs, err := a.cache.Get(rt)
 	if err != nil {
 		if errors.Cause(err) != errcode.ErrCacheKeyNotFound {
 			return nil, errors.WithStack(err)
 		}
 
-		rs, err = iamUsers(ctx, a, rt, tags)
+		rs, err = iamUsers(ctx, a, rt, filters)
 		if err != nil {
 			return nil, err
 		}
@@ -178,8 +178,8 @@ func cacheIAMUsers(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]pr
 	return rs, nil
 }
 
-func getIAMUserNames(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]string, error) {
-	rs, err := cacheIAMUsers(ctx, a, rt, tags)
+func getIAMUserNames(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]string, error) {
+	rs, err := cacheIAMUsers(ctx, a, rt, filters)
 	if err != nil {
 		return nil, err
 	}
@@ -194,14 +194,14 @@ func getIAMUserNames(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]
 	return names, nil
 }
 
-func cacheRoute53Zones(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]provider.Resource, error) {
+func cacheRoute53Zones(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]provider.Resource, error) {
 	rs, err := a.cache.Get(rt)
 	if err != nil {
 		if errors.Cause(err) != errcode.ErrCacheKeyNotFound {
 			return nil, errors.WithStack(err)
 		}
 
-		rs, err = route53Zones(ctx, a, rt, tags)
+		rs, err = route53Zones(ctx, a, rt, filters)
 		if err != nil {
 			return nil, err
 		}
@@ -215,8 +215,8 @@ func cacheRoute53Zones(ctx context.Context, a *aws, rt string, tags []tag.Tag) (
 	return rs, nil
 }
 
-func getRoute53ZoneIDs(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]string, error) {
-	rs, err := cacheRoute53Zones(ctx, a, rt, tags)
+func getRoute53ZoneIDs(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]string, error) {
+	rs, err := cacheRoute53Zones(ctx, a, rt, filters)
 	if err != nil {
 		return nil, err
 	}
@@ -231,14 +231,14 @@ func getRoute53ZoneIDs(ctx context.Context, a *aws, rt string, tags []tag.Tag) (
 	return ids, nil
 }
 
-func cacheSESDomainIdentities(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]provider.Resource, error) {
+func cacheSESDomainIdentities(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]provider.Resource, error) {
 	rs, err := a.cache.Get(rt)
 	if err != nil {
 		if errors.Cause(err) != errcode.ErrCacheKeyNotFound {
 			return nil, errors.WithStack(err)
 		}
 
-		rs, err = sesDomainIdentities(ctx, a, rt, tags)
+		rs, err = sesDomainIdentities(ctx, a, rt, filters)
 		if err != nil {
 			return nil, err
 		}
@@ -252,8 +252,8 @@ func cacheSESDomainIdentities(ctx context.Context, a *aws, rt string, tags []tag
 	return rs, nil
 }
 
-func getSESDomainIdentityDomains(ctx context.Context, a *aws, rt string, tags []tag.Tag) ([]string, error) {
-	rs, err := cacheSESDomainIdentities(ctx, a, rt, tags)
+func getSESDomainIdentityDomains(ctx context.Context, a *aws, rt string, filters *filter.Filter) ([]string, error) {
+	rs, err := cacheSESDomainIdentities(ctx, a, rt, filters)
 	if err != nil {
 		return nil, err
 	}