+++ aws/provider.go
@@ -69,7 +69,7 @@ func (a *aws) Resources(ctx context.Context, t string, f *filter.Filter) ([]prov
 		return nil, errors.Errorf("the resource %q it's not implemented", t)
 	}
 
-	resources, err := rfn(ctx, a, t, f.Tags)
+	resources, err := rfn(ctx, a, t, f)
 	if err != nil {
 		return nil, errors.Wrapf(err, "error while reading from resource %q", t)
 	}