+++ azurerm/provider.go
@@ -84,7 +84,7 @@ func (a *azurerm) Resources(ctx context.Context, t string, f *filter.Filter) ([]
 		return nil, errors.Errorf("the resource %q it's not implemented", t)
 	}
 
-	resources, err := rfn(ctx, a, t, f.Tags)
+	resources, err := rfn(ctx, a, t, f)
 	if err != nil {
 		return nil, errors.Wrapf(err, "error while reading from resource %q", t)
 	}